adds loggedInLables in header as well as loggedOutLables. Renames component to HeaderMain. Creates separate component file for HeaderCategories. Downloads BrowserRouter to prepare for creating navlinks
