+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^0.87.1"
   },
