+++ src/index.js
@@ -3,7 +3,12 @@ import ReactDOM from 'react-dom';
 import 'semantic-ui-css/semantic.min.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
+import { Switch, Route, BrowserRouter } from 'react-router-dom'
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render((
+<BrowserRouter>
+  <App />
+</BrowserRouter>
+), document.getElementById('root'));
 
 serviceWorker.unregister();
